feat: user history

Closes S05P21A202-161

See merge request s05-bigdata-dist/S05P21A202!15

+++ backend/gumison/src/main/java/com/ssafy/gumison/db/entity/LevelTier.java
@@ -30,8 +30,8 @@ public class LevelTier {
 	private Climbing climbing;
 
 	@Column(nullable = false)
-	private Integer levelCode;
+	private Long levelCode;
 
 	@Column(nullable = false)
-	private Integer tierCode;
+	private Long tierCode;
 }
